+++ storm/src/test/java/com/hortonworks/iotas/rules/GroovyTest.java
@@ -108,4 +108,5 @@ public class GroovyTest {
 
         script.run();  // throws exception because no bindings for x, y
     }
+
 }